+++ R/decimal_truth.R
@@ -14,12 +14,12 @@
 #'  \code{sqrt(.Machine$double.eps)},
 #'  then \code{x} and \code{y}
 #'  should be consider to be equal. \cr
-#'  For example: \code{0.1*7 == 0.7} returns \code{FALSE}, even though they are equal,
+#'  For example: \code{(0.1 * 7) == 0.7} returns \code{FALSE}, even though they are equal,
 #'  due to the way decimal numbers are stored in programming languages like 'R' and  'Python'. \cr
-#'  But \code{0.1*7 %d==% 0.7} returns \code{TRUE}.
+#'  But \code{(0.1 * 7) %d==% 0.7} returns \code{TRUE}. \cr
 #'  2) Only numeric input is allowed, so characters are not coerced to numbers. \cr
 #'  I.e. \code{1 < "a"} gives \code{TRUE}, whereas \code{1 %d<% "a"} gives an error. \cr
-#'  For character equality testing, see \link[stringi]{%s==%} from the 'stringi' package.
+#'  For character equality testing, see \link[stringi]{%s==%} from the 'stringi' package. \cr
 #' 
 #' Thus these operators provide safer decimal number (in)equality tests. \cr
 #' \cr
@@ -42,63 +42,21 @@
 #' @param tol a single, strictly positive number close to zero, giving the tolerance.
 #'
 #' @returns
+#' For the \code{%d...%} operators: \cr
 #' A logical vector with the same dimensions as \code{x},
-#' indicating the result of the element by element comparison.
+#' indicating the result of the element by element comparison. \cr
+#' NOTE: `Inf` by `Inf` and `-Inf` by `-Inf` comparisons with
+#' the \code{%d...%} operators return `NA`. \cr
+#' \cr
+#' For `is_wholenumber()`: \cr
+#' A logical vector with the same dimensions as \code{x},
+#' indicating the result of the element by element comparison. \cr
+#' NOTE: `Inf`, `-Inf`, `NaN` and `NA` all return `NA` for `is_wholenumber()`.
 #'
 #' @seealso \link{tinycodet_safer}
 #'
-#' @examples
-#' x <- c(0.3, 0.6, 0.7)
-#' y <- c(0.1*3, 0.1*6, 0.1*7)
-#' print(x); print(y)
-#' x == y # gives FALSE, but should be TRUE
-#' x!= y # gives TRUE, should be FALSE
-#' x > y # not wrong
-#' x < y # gives TRUE, should be FALSE
-#' x %d==% y # here it's done correctly
-#' x %d!=% y # correct
-#' x %d<% y # correct
-#' x %d>% y # correct
-#' x %d<=% y # correct
-#' x %d>=% y # correct
-#'
-#' x <- c(0.3, 0.6, 0.7)
-#' bnd <- cbind(x-0.1, x+0.1)
-#' x %d{}% bnd
-#' x %d!{}% bnd
-#'
-#' # These operators work for integers also:
-#' x <- 1L:5L
-#' y <- 1L:5L
-#' x %d==% y
-#' x %d!=% y
-#' x %d<% y
-#' x %d>% y
-#' x %d<=% y
-#' x %d>=% y
-#'
-#' x <- 1L:5L
-#' y <- x+1
-#' x %d==% y
-#' x %d!=% y
-#' x %d<% y
-#' x %d>% y
-#' x %d<=% y
-#' x %d>=% y
-#'
-#' x <- 1L:5L
-#' y <- x-1
-#' x %d==% y
-#' x %d!=% y
-#' x %d<% y
-#' x %d>% y
-#' x %d<=% y
-#' x %d>=% y
-#' 
-#' # is_wholenumber:
-#' is_wholenumber(1:10 + c(0, 0.1))
-#' 
-#' 
+#' @example inst/examples/decimal_truth.R
+
 
 #' @name decimal_truth
 NULL
@@ -120,7 +78,7 @@ NULL
 #' @rdname decimal_truth
 #' @export
 `%d<%` <- function(x, y) {
-  return((y - x) >= sqrt(.Machine$double.eps))
+  return((x - y) <= - sqrt(.Machine$double.eps))
 }
 
 #' @rdname decimal_truth
@@ -132,15 +90,15 @@ NULL
 #' @rdname decimal_truth
 #' @export
 `%d<=%` <- function(x, y) {
-  
-  return((x <= y) | (x %d==% y))
+  out <- (x - y) < sqrt(.Machine$double.eps) 
+  return(out)
 }
 
 #' @rdname decimal_truth
 #' @export
 `%d>=%` <- function(x, y) {
-  
-  return((x >= y) | (x %d==% y))
+  out <- (x - y) > -sqrt(.Machine$double.eps) 
+  return(out)
 }
 
 #' @rdname decimal_truth