+++ inst/tinytest/safer/test-decimal-eq.R
@@ -2,6 +2,8 @@
 # set-up ====
 tol <- sqrt(.Machine$double.eps)
 eps <- tol * 2
+enumerate <- 0 # to count number of tests performed using iterations in loops
+loops <- 0 # to count number of loops
 
 # basic checks ====
 
@@ -24,7 +26,6 @@ equal <- c(rep(TRUE, 3), rep(FALSE, 6), rep(NA, 6))
 smaller <- c(rep(FALSE, 6), rep(TRUE, 3), rep(NA, 6))
 bigger <- c(rep(FALSE, 3), rep(TRUE, 3), rep(FALSE, 3), rep(NA, 6))
 
-# regular checks
 expect_equal(x %d==% y, equal)
 expect_equal(x %d<=% y, equal | smaller)
 expect_equal(x %d>=% y, equal | bigger)
@@ -32,10 +33,12 @@ expect_equal(x %d!=% y, !equal)
 expect_equal(x %d<% y, !equal & smaller)
 expect_equal(x %d>% y, !equal & bigger)
 
-# relational checks
+
+# relational checks ====
 expect_equal(x %d!=% y, !(x %d==% y))
 expect_equal(x %d<=% y, !(x %d>% y))
 expect_equal(x %d>=% y, !(x %d<% y))
+
 
 
 # recycling checks ====
@@ -71,13 +74,74 @@ expect_error(x %d<% ybad, pattern = "non-conformable arrays")
 expect_error(x %d>=% ybad, pattern = "non-conformable arrays")
 expect_error(x %d<=% ybad, pattern = "non-conformable arrays")
 
-# errors ====
-expect_error(1 %d==% "a")
-expect_error(1 %d!=% "a")
-expect_error(1 %d<=% "a")
-expect_error(1 %d>=% "a")
-expect_error(1 %d<% "a")
-expect_error(1 %d>% "a")
-expect_error(1 %d{}% "a")
-expect_error(1 %d!{}% "a")
+
+# attribute checks ====
+x.data <- list(
+  sample(1:20),
+  structure(
+    sample(1:20),
+    dim = c(5, 4),
+    dimnames = list(month.abb[1:5], month.abb[1:4]), names = letters[1:20],
+    test = "test1"
+  ),
+  structure(
+    sample(1:20),
+    dim = c(5, 4),
+    dimnames = list(month.name[1:5], month.name[1:4]), names = LETTERS[1:20],
+    test = "test2"
+  )
+)
+y.data <- list(
+  sample(1:20),
+  structure(
+    sample(1:20),
+    dim = c(5, 4),
+    dimnames = list(month.abb[1:5], month.abb[1:4]), names = letters[1:20],
+    test = "test1"
+  ),
+  structure(
+    sample(1:20),
+    dim = c(5, 4),
+    dimnames = list(month.name[1:5], month.name[1:4]), names = LETTERS[1:20],
+    test = "test2"
+  )
+)
+errorfun <- function(tt) {
+  if(isTRUE(tt)) print(tt)
+  if(isFALSE(tt)) stop(print(tt))
+}
+loops <- loops + 1
+for(i in seq_along(x.data)) {
+  for(j in seq_along(y.data)) {
+    x <- x.data[[i]]
+    y <- x.data[[j]]
+    
+    expect_equal(
+      x == y,
+      x %d==% y
+    ) |> errorfun()
+    expect_equal(
+      x != y,
+      x %d!=% y
+    ) |> errorfun()
+    expect_equal(
+      x <= y,
+      x %d<=% y
+    ) |> errorfun()
+    expect_equal(
+      x >= y,
+      x %d>=% y
+    ) |> errorfun()
+    expect_equal(
+      x < y,
+      x %d<% y
+    ) |> errorfun()
+    expect_equal(
+      x > y,
+      x %d>% y
+    ) |> errorfun()
+    
+    enumerate <- enumerate + 6
+  }
+}