+++ inst/tinytest/special/test-import_as-special.R
@@ -10,7 +10,6 @@ print(lib.loc1)
 print(lib.loc2)
 print(lib.loc3)
 
-
 # test import_as - single package ====
 stri <- loadNamespace("stringi") |> getNamespaceExports()
 import_as(~stri., "stringi")
@@ -20,13 +19,13 @@ expect_equal(out, sort(stri))
 
 # incorrect package error handling ====
 expect_error(
-  import_as(~p1., "tinycodetfakepkg1", lib.loc=c(lib.loc1, .libPaths()), dependencies = "stringi"),
+  import_as(~p1., "tinycodetfakepkg1", lib.loc = c("foo1", lib.loc1, .libPaths(), "foo2"), dependencies = "stringi"),
   pattern = "The following given dependencies were not found to be actual dependencies"
 )
 
 
 expect_error(
-  import_as(~p1., "tinycodetfakepkg1", lib.loc=c(lib.loc1, .libPaths()), extensions = "stringi"),
+  import_as(~p1., "tinycodetfakepkg1", lib.loc = c("foo1", lib.loc1, .libPaths(), "foo2"), extensions = "stringi"),
   pattern = "The following given extensions were not found to be actual extensions"
 )
 
@@ -34,7 +33,7 @@ expect_error(
 # duplicate package error handling ====
 dupli_pkg <- "stringi"
 expect_error(
-  import_as(~p1., "tinycodetfakepkg1", lib.loc=c(lib.loc1, .libPaths()),
+  import_as(~p1., "tinycodetfakepkg1", lib.loc = c("foo1", lib.loc1, .libPaths(), "foo2"),
             dependencies = rep(dupli_pkg, 2)),
   pattern = paste0("The following duplicate dependencies given:", "\n",
                    dupli_pkg)
@@ -42,7 +41,7 @@ expect_error(
 
 
 expect_error(
-  import_as(~p1., "tinycodetfakepkg1", lib.loc=c(lib.loc1, .libPaths()),
+  import_as(~p1., "tinycodetfakepkg1", lib.loc = c("foo1", lib.loc1, .libPaths(), "foo2"),
             extensions = rep(dupli_pkg, 2)),
   pattern = paste0("The following duplicate extensions given:", "\n",
                    dupli_pkg)
@@ -50,7 +49,7 @@ expect_error(
 
 dupli_pkg <- "tidytable"
 expect_error(
-  import_as(~p1., "tinycodetfakepkg1", lib.loc=c(lib.loc1, .libPaths()),
+  import_as(~p1., "tinycodetfakepkg1", lib.loc = c("foo1", lib.loc1, .libPaths(), "foo2"),
             dependencies = rep(dupli_pkg, 2)),
   pattern = paste0("The following duplicate dependencies given:", "\n",
                    dupli_pkg)
@@ -58,7 +57,7 @@ expect_error(
 
 
 expect_error(
-  import_as(~p1., "tinycodetfakepkg1", lib.loc=c(lib.loc1, .libPaths()),
+  import_as(~p1., "tinycodetfakepkg1", lib.loc = c("foo1", lib.loc1, .libPaths(), "foo2"),
             extensions = rep(dupli_pkg, 2)),
   pattern = paste0("The following duplicate extensions given:", "\n",
                    dupli_pkg)
@@ -67,25 +66,25 @@ expect_error(
 
 # missing package error handling ====
 expect_error(
-  import_as(~ p3., "tinycodetfakepkg3", dependencies = "tinycodetfakepkg1", lib.loc=lib.loc2),
+  import_as(~ p3., "tinycodetfakepkg3", dependencies = "tinycodetfakepkg1", lib.loc = c("foo1", lib.loc2, "foo2")),
   pattern = "The following dependencies are not installed",
   fixed = TRUE
 )
 
 expect_error(
-  import_as(~ p3., "tinycodetfakepkg3", dependencies = "tinycodetfakepkg2", lib.loc=lib.loc2),
+  import_as(~ p3., "tinycodetfakepkg3", dependencies = "tinycodetfakepkg2", lib.loc = c("foo1", lib.loc2, "foo2")),
   pattern = "The following dependencies are not installed",
   fixed = TRUE
 )
 
 expect_error(
-  import_as(~p1., "tinycodetfakepkg1", extensions = "tinycodetfakepkg3", lib.loc=lib.loc3),
+  import_as(~p1., "tinycodetfakepkg1", extensions = "tinycodetfakepkg3", lib.loc = c("foo1", lib.loc3, "foo2")),
   pattern = "The following extensions are not installed",
   fixed = TRUE
 )
 
 expect_error(
-  import_as(~p2., "tinycodetfakepkg2", extensions = "tinycodetfakepkg3", lib.loc=lib.loc3),
+  import_as(~p2., "tinycodetfakepkg2", extensions = "tinycodetfakepkg3", lib.loc = c("foo1", lib.loc3, "foo2")),
   pattern = "The following extensions are not installed",
   fixed = TRUE
 )
@@ -95,7 +94,7 @@ expect_error(
 import_as(
   ~p3., "tinycodetfakepkg3",
   re_exports=TRUE,
-  lib.loc = lib.loc1
+  lib.loc = c("foo1", lib.loc1, "foo2")
 )
 p3 <- c(
   "fun_overwritten", "%opover%",
@@ -106,8 +105,8 @@ out <- setdiff(names(p3.), ".__attributes__.") |> sort()
 expect_equal(out,  sort(p3))
 expect_true(p3.$.__attributes__.$args$re_exports)
 expect_equal(
-  import_as(~ p3., "tinycodetfakepkg3", re_exports = FALSE, dependencies = "tinycodetfakepkg1", lib.loc = lib.loc1),
-  import_as(~ p3., "tinycodetfakepkg1", re_exports = FALSE, extensions = "tinycodetfakepkg3", lib.loc = lib.loc1),
+  import_as(~ p3., "tinycodetfakepkg3", re_exports = FALSE, dependencies = "tinycodetfakepkg1", lib.loc = c("foo1", lib.loc1, "foo2")),
+  import_as(~ p3., "tinycodetfakepkg1", re_exports = FALSE, extensions = "tinycodetfakepkg3", lib.loc = c("foo1", lib.loc1, "foo2")),
 )
 
 
@@ -116,7 +115,7 @@ import_as(
   ~p3., "tinycodetfakepkg3",
   re_exports = TRUE,
   dependencies=c("tinycodetfakepkg1", "tinycodetfakepkg2"),
-  lib.loc=lib.loc1
+  lib.loc = c("foo1", lib.loc1, "foo2")
 )
 p3 <- c(
   "fun_overwritten", "%opover%",
@@ -143,7 +142,7 @@ expect_equal(
 import_as(
   ~ p3., "tinycodetfakepkg1",
   extensions = "tinycodetfakepkg3",
-  lib.loc=lib.loc1
+  lib.loc = c("foo1", lib.loc1, "foo2")
 )
 p3 <- c(
   "fun_overwritten", "%opover%",
@@ -170,7 +169,7 @@ import_as(
   ~ p3., "tinycodetfakepkg3",
   re_exports = TRUE,
   dependencies=c("tinycodetfakepkg1", "tinycodetfakepkg2"),
-  lib.loc=lib.loc1
+  lib.loc = c("foo1", lib.loc1, "foo2")
 ) |> suppressMessages()
 p3 <- data.frame(
   package=c("tinycodetfakepkg1", "tinycodetfakepkg2", "tinycodetfakepkg3 + re-exports"),
@@ -184,7 +183,7 @@ import_as(
   ~ p3., "tinycodetfakepkg3",
   re_exports = FALSE,
   dependencies=c("tinycodetfakepkg1", "tinycodetfakepkg2"),
-  lib.loc=lib.loc1
+  lib.loc = c("foo1", lib.loc1, "foo2")
 ) |> suppressMessages()
 p3 <- data.frame(
   package=c("tinycodetfakepkg1", "tinycodetfakepkg2", "tinycodetfakepkg3"),
@@ -198,7 +197,7 @@ import_as(
   ~ p3., "tinycodetfakepkg3",
   re_exports = TRUE,
   dependencies=c("tinycodetfakepkg1", "tinycodetfakepkg2"),
-  lib.loc=lib.loc1,
+  lib.loc = c("foo1", lib.loc1, "foo2"),
   import_order = c("main_package", "dependencies", "extensions")
 ) |> suppressMessages()
 p3 <- data.frame(
@@ -215,7 +214,7 @@ expect_equal(
 expect_silent(import_as(
   ~ new., "tinycodetfakepkg3",
   re_exports = TRUE,
-  lib.loc=lib.loc1
+  lib.loc = c("foo1", lib.loc1, "foo2")
 ) |> suppressMessages())
 
 
@@ -225,7 +224,7 @@ import_as(
   re_exports = FALSE,
   dependencies = c("tinycodetfakepkg2", "tinycodetfakepkg1"),
   import_order = c("main_package", "dependencies", "extensions"),
-  lib.loc = lib.loc1
+  lib.loc = c("foo1", lib.loc1, "foo2")
 )  |> suppressMessages()
 expect_equal(
   new.$.__attributes__.$pkgs$packages_order,
@@ -233,7 +232,7 @@ expect_equal(
 )
 ordered_object_names <- sapply(
   c("tinycodetfakepkg3", "tinycodetfakepkg2",  "tinycodetfakepkg1"),
-  \(x)pkg_lsf(x, type = "all", lib.loc = lib.loc1)
+  \(x)pkg_lsf(x, type = "all", lib.loc = c("foo1", lib.loc1, "foo2"))
 ) |> as.character() |> unique()
 expect_equal(
   new.$.__attributes__.$ordered_object_names,
@@ -245,7 +244,7 @@ import_as(
   re_exports = FALSE,
   extensions = c("tinycodetfakepkg3"),
   import_order = c("extensions", "main_package", "dependencies"),
-  lib.loc = lib.loc1
+  lib.loc = c("foo1", lib.loc1, "foo2")
 ) |> suppressMessages()
 expect_equal(
   new.$.__attributes__.$pkgs$packages_order,
@@ -253,7 +252,7 @@ expect_equal(
 )
 ordered_object_names <- sapply(
   c("tinycodetfakepkg3", "tinycodetfakepkg1"),
-  \(x)pkg_lsf(x, type = "all", lib.loc = lib.loc1)
+  \(x)pkg_lsf(x, type = "all", lib.loc = c("foo1", lib.loc1, "foo2"))
 ) |> as.character() |> unique()
 expect_equal(
   new.$.__attributes__.$ordered_object_names,
@@ -264,7 +263,7 @@ import_as(
   ~ new., "tinycodetfakepkg1",
   re_exports = FALSE,
   import_order = c("main_package", "dependencies", "extensions"),
-  lib.loc = lib.loc1
+  lib.loc = c("foo1", lib.loc1, "foo2")
 ) |> suppressMessages()
 expect_equal(
   new.$.__attributes__.$pkgs$packages_order,
@@ -272,7 +271,7 @@ expect_equal(
 )
 ordered_object_names <- sapply(
   c("tinycodetfakepkg1"),
-  \(x)pkg_lsf(x, type = "all", lib.loc = lib.loc1)
+  \(x)pkg_lsf(x, type = "all", lib.loc = c("foo1", lib.loc1, "foo2"))
 ) |> as.character() |> unique()
 expect_equal(
   new.$.__attributes__.$ordered_object_names,
@@ -284,7 +283,7 @@ expect_equal(
 import_as(
   ~ new., "tinycodetfakepkg3",
   re_exports = TRUE,
-  lib.loc=lib.loc1
+  lib.loc = c("foo1", lib.loc1, "foo2")
 )  |> suppressMessages()
 expect_true("tinyimport" %in% names(new.$.__attributes__.))