+++ inst/tinytest/special/test-import_general-special.R
@@ -12,20 +12,36 @@ print(lib.loc3)
 
 
 # functional base functions from separate library ====
-import_as(~new., "tinycodetfakepkg5", lib.loc = lib.loc1)
+# single lib.loc
+import_as(~new., "tinycodetfakepkg5", lib.loc = c("foo1", lib.loc1, "foo2"))
 expect_equal(new.$fun_paste("a", "b"), "ab")
 
 temp.fun <- function() {
-  import_inops("tinycodetfakepkg5", lib.loc = lib.loc1)
+  import_inops("tinycodetfakepkg5", lib.loc = c("foo1", lib.loc1, "foo2"))
   "a" %paste0% "b"
 }
 expect_equal(temp.fun(), "ab")
 temp.fun <- function() {
-  import_LL("tinycodetfakepkg5", "fun_paste", lib.loc = lib.loc1)
+  import_LL("tinycodetfakepkg5", "fun_paste", lib.loc = c("foo1", lib.loc1, "foo2"))
   fun_paste("a", "b")
 }
 expect_equal(temp.fun(), "ab")
 
+
+# multi lib.loc
+import_as(~new., "tinycodetfakepkg5", lib.loc = c("foo", lib.loc1))
+expect_equal(new.$fun_paste("a", "b"), "ab")
+
+temp.fun <- function() {
+  import_inops("tinycodetfakepkg5", lib.loc = c("foo", lib.loc1))
+  "a" %paste0% "b"
+}
+expect_equal(temp.fun(), "ab")
+temp.fun <- function() {
+  import_LL("tinycodetfakepkg5", "fun_paste", lib.loc = c("foo", lib.loc1))
+  fun_paste("a", "b")
+}
+expect_equal(temp.fun(), "ab")
 
 
 
@@ -35,45 +51,77 @@ pattern <- paste0(
   "tinycodetfakepkg3",
   "`, the following packages are required but not installed:"
 )
+
+# single lib.loc
 expect_error(
-  import_as(~ p3., "tinycodetfakepkg3", re_exports=TRUE, lib.loc=lib.loc2),
+  import_as(~ p3., "tinycodetfakepkg3", re_exports=TRUE, lib.loc = c("foo1", lib.loc2, "foo2")),
   pattern = pattern,
   fixed = TRUE
 )
 
 expect_error(
-  import_as(~ p3., "tinycodetfakepkg3", re_exports=FALSE, lib.loc=lib.loc2),
+  import_as(~ p3., "tinycodetfakepkg3", re_exports=FALSE, lib.loc = c("foo1", lib.loc2, "foo2")),
   pattern = pattern,
   fixed = TRUE
 )
 
 expect_error(
-  import_as(~ p3., "tinycodetfakepkg3", re_exports=TRUE, lib.loc=lib.loc2),
+  import_as(~ p3., "tinycodetfakepkg3", re_exports=TRUE, lib.loc = c("foo1", lib.loc2, "foo2")),
   pattern = pattern,
   fixed = TRUE
 )
 
 expect_error(
-  import_inops("tinycodetfakepkg3", lib.loc=lib.loc2),
+  import_inops("tinycodetfakepkg3", lib.loc = c("foo1", lib.loc2, "foo2")),
   pattern = pattern,
   fixed = TRUE
 )
 
 expect_error(
-  import_LL("tinycodetfakepkg3", "foo", lib.loc=lib.loc2),
+  import_LL("tinycodetfakepkg3", "foo", lib.loc = c("foo1", lib.loc2, "foo2")),
+  pattern = pattern,
+  fixed = TRUE
+)
+
+# multi lib.loc
+expect_error(
+  import_as(~ p3., "tinycodetfakepkg3", re_exports=TRUE, lib.loc = c("foo", lib.loc2)),
+  pattern = pattern,
+  fixed = TRUE
+)
+
+expect_error(
+  import_as(~ p3., "tinycodetfakepkg3", re_exports=FALSE, lib.loc = c("foo", lib.loc2)),
+  pattern = pattern,
+  fixed = TRUE
+)
+
+expect_error(
+  import_as(~ p3., "tinycodetfakepkg3", re_exports=TRUE, lib.loc = c("foo", lib.loc2)),
+  pattern = pattern,
+  fixed = TRUE
+)
+
+expect_error(
+  import_inops("tinycodetfakepkg3", lib.loc = c("foo", lib.loc2)),
+  pattern = pattern,
+  fixed = TRUE
+)
+
+expect_error(
+  import_LL("tinycodetfakepkg3", "foo", lib.loc = c("foo", lib.loc2)),
   pattern = pattern,
   fixed = TRUE
 )
 
 
-# function attributes ====
+# function attributes - import_as ====
 
-# import_as
 import_as(
   ~ p3., "tinycodetfakepkg3",
   re_exports = FALSE,
   dependencies = c("tinycodetfakepkg1", "tinycodetfakepkg2"),
-  lib.loc = lib.loc1
+  lib.loc = c("foo1", lib.loc1, "foo2")
 )
 checklist <- as.list(p3.)
 checklist <- checklist[order(names(checklist))]
@@ -88,14 +136,13 @@ expect_equal(
 )
 
 
-# import_inops
-
+# function attributes - import_inops ====
 tempfun <- function() {
   import_as(
     ~ p3., "tinycodetfakepkg3",
     re_exports = FALSE,
     dependencies = c("tinycodetfakepkg1", "tinycodetfakepkg2"),
-    lib.loc = lib.loc1
+    lib.loc = c("foo1", lib.loc1, "foo2")
   )
   import_inops(p3.)
   funnames <- setdiff(ls(), "p3.")
@@ -114,11 +161,12 @@ out <- tempfun() |> unname()
 expect_equal(expect, out)
 
 
-# import_LL
+# function attributes - import_LL ====
+
 tempfun <- function() {
-  import_LL("tinycodetfakepkg1", "fun11", lib.loc = lib.loc1)
-  import_LL("tinycodetfakepkg2", "fun21", lib.loc = lib.loc1)
-  import_LL("tinycodetfakepkg3", "fun31", lib.loc = lib.loc1)
+  import_LL("tinycodetfakepkg1", "fun11", lib.loc = c("foo1", lib.loc1, "foo2"))
+  import_LL("tinycodetfakepkg2", "fun21", lib.loc = c("foo1", lib.loc1, "foo2"))
+  import_LL("tinycodetfakepkg3", "fun31", lib.loc = c("foo1", lib.loc1, "foo2"))
   out <- c(attr(fun11, "package"), attr(fun21, "package"), attr(fun31, "package"))
   return(out)
 }
@@ -128,7 +176,6 @@ expect <- c("tinycodetfakepkg1", "tinycodetfakepkg2", "tinycodetfakepkg3")
 out <- tempfun()
 
 expect_equal(expect, out)
-